+++ src/modules/searchBar/tests/unit/searchInterfaces/searchInterface.spec.js
@@ -190,4 +190,51 @@ describe("src/modules/searchBar/searchInterfaces/searchInterface.js", () => {
             });
         });
     });
+
+    describe("normalizeResultEvents", () => {
+        let resultEventsToObjectSpy,
+            createPossibleActionsSpy,
+            resultEvents,
+            resultEventsAsObject,
+            actions;
+
+        beforeEach(() => {
+            resultEvents = {
+                onClick: ["setMarker", "zoomToResult"]
+            };
+            resultEventsAsObject = {
+                onClick: {
+                    setMarker: {coordinates: [1, 2]},
+                    zoomToResult: {coordinates: [1, 2]}
+                }
+            };
+            actions = {
+                setMarker: {coordinates: [1, 2]},
+                zoomToResult: {coordinates: [1, 2]}
+            };
+            resultEventsToObjectSpy = sinon.stub(SearchInterface.prototype, "resultEventsToObject").returns(resultEventsAsObject);
+            createPossibleActionsSpy = sinon.stub(SearchInterface.prototype, "createPossibleActions").returns(actions);
+        });
+
+        it("all available", () => {
+            expect(SearchInterface1.normalizeResultEvents(resultEvents)).to.deep.equals({
+                onClick: {
+                    setMarker: {coordinates: [1, 2]},
+                    zoomToResult: {coordinates: [1, 2]}
+                }
+            });
+            expect(resultEventsToObjectSpy.calledOnce).to.be.true;
+            expect(createPossibleActionsSpy.calledOnce).to.be.true;
+        });
+        it("one action not available", () => {
+            delete actions.setMarker;
+            expect(SearchInterface1.normalizeResultEvents(resultEvents)).to.deep.equals({
+                onClick: {
+                    zoomToResult: {coordinates: [1, 2]}
+                }
+            });
+            expect(resultEventsToObjectSpy.calledOnce).to.be.true;
+            expect(createPossibleActionsSpy.calledOnce).to.be.true;
+        });
+    });
 });