+++ src/modules/searchBar/tests/unit/searchInterfaces/searchInterfaceElasticSearch.spec.js
@@ -234,10 +234,13 @@ describe("src/modules/searchBar/searchInterfaces/searchInterfaceElasticSearch.js
     });
 
     describe("normalizeResults", () => {
-        it("should normalize an search result", () => {
-            const searchResults = [
+        let searchResults,
+            result;
+
+        beforeEach(() => {
+            searchResults = [
                 {
-                    _id: "100",
+                    _id: 100,
                     _source: {
                         id: "123",
                         name: "Test abc",
@@ -248,8 +251,7 @@ describe("src/modules/searchBar/searchInterfaces/searchInterfaceElasticSearch.js
                         ]
                     }
                 }];
-
-            expect(SearchInterface1.normalizeResults(searchResults)).to.deep.equals([
+            result = [
                 {
                     events: {
                         onClick: {
@@ -299,13 +301,28 @@ describe("src/modules/searchBar/searchInterfaces/searchInterfaceElasticSearch.js
                     name: "Test abc",
                     toolTip: "Test abc - md name"
                 }
-            ]);
+            ];
+        });
+
+        it("should normalize an search result", () => {
+            expect(SearchInterface1.normalizeResults(searchResults)).to.deep.equals(result);
+        });
+
+        it("should normalize an search result with id of type number, besides in source", () => {
+            searchResults[0]._source.id = 123;
+            result[0].events.onClick.addLayerToTopicTree.source.id = 123;
+            result[0].events.buttons.showInTree.source.id = 123;
+            result[0].events.buttons.showLayerInfo.source.id = 123;
+            expect(SearchInterface1.normalizeResults(searchResults)).to.deep.equals(result);
         });
     });
 
     describe("createPossibleActions", () => {
-        it("should create possible events from search result", () => {
-            const searchResult = {
+        let searchResult,
+            result;
+
+        beforeEach(() => {
+            searchResult = {
                 _id: "100",
                 _source: {
                     id: "123",
@@ -316,33 +333,37 @@ describe("src/modules/searchBar/searchInterfaces/searchInterfaceElasticSearch.js
                     }
                 }
             };
-
-            expect(SearchInterface1.createPossibleActions(searchResult)).to.deep.equals(
-                {
-                    addLayerToTopicTree: {
-                        layerId: searchResult._source.id,
-                        source: searchResult._source
-                    },
-                    setMarker: {
-                        coordinates: [10, 20]
-                    },
-                    zoomToResult: {
-                        coordinates: [10, 20]
-                    },
-                    startRouting: {
-                        coordinates: [10, 20],
-                        name: searchResult._source.name
-                    },
-                    showInTree: {
-                        layerId: searchResult._source.id,
-                        source: searchResult._source
-                    },
-                    showLayerInfo: {
-                        layerId: searchResult._source.id,
-                        source: searchResult._source
-                    }
+            result = {
+                addLayerToTopicTree: {
+                    layerId: searchResult._source.id,
+                    source: searchResult._source
+                },
+                setMarker: {
+                    coordinates: [10, 20]
+                },
+                zoomToResult: {
+                    coordinates: [10, 20]
+                },
+                startRouting: {
+                    coordinates: [10, 20],
+                    name: searchResult._source.name
+                },
+                showInTree: {
+                    layerId: searchResult._source.id,
+                    source: searchResult._source
+                },
+                showLayerInfo: {
+                    layerId: searchResult._source.id,
+                    source: searchResult._source
                 }
-            );
+            };
+        });
+        it("should create possible events from search result", () => {
+            expect(SearchInterface1.createPossibleActions(searchResult)).to.deep.equals(result);
+        });
+        it("should create possible events from search result, id as number", () => {
+            searchResult._source.id = 123;
+            expect(SearchInterface1.createPossibleActions(searchResult)).to.deep.equals(result);
         });
     });